renamed DiagnosticStatus to DiagnosticTable
