+++ packages/suite-base/src/panels/diagnostics/DiagnosticTable.style.ts
